+++ lib/puree/extractor/project.rb
@@ -9,7 +9,7 @@ module Puree
       # @option (see Puree::Extractor::Resource#initialize)
       def initialize(config)
         super
-        @latest_api = false # stable API does not return person roles
+        @latest_api = false # current API does not return person roles
         setup :project
       end
 
@@ -18,7 +18,10 @@ module Puree
       def combine_metadata
         super
         @model.acronym = @extractor.acronym
+        # @model.associated = @extractor.associated # not present in stable API
         @model.description = @extractor.description
+        @model.external_organisations = @extractor.external_organisations
+        @model.funded = @extractor.funded?
         @model.organisations = @extractor.organisations
         @model.owner = @extractor.owner
         @model.persons_internal = @extractor.persons_internal