+++ lib/puree/model/project.rb
@@ -8,8 +8,18 @@ module Puree
       # @return [String, nil]
       attr_accessor :acronym
 
+      # not present in stable API
+      # @return [Array<Puree::Model::RelatedContentHeader>]
+      # attr_accessor :associated
+
       # @return [String, nil]
       attr_accessor :description
+
+      # @return [Array<Puree::Model::ExternalOrganisationHeader>]
+      attr_accessor :external_organisations
+
+      # @return [Boolean]
+      attr_accessor :funded
 
       # @return [Array<Puree::Model::OrganisationHeader>]
       attr_accessor :organisations