+++ lib/puree/xml_extractor/publication.rb
@@ -5,6 +5,8 @@ module Puree
     # Publication XML extractor.
     #
     class Publication < Puree::XMLExtractor::Resource
+      include Puree::XMLExtractor::AssociatedMixin
+      include Puree::XMLExtractor::ExternalOrganisationsMixin
 
       def initialize(xml:)
         @resource_type = :publication